Post review refactoring v2 * Gets rid of BasicSttpClientXYZ and uses SttpClientXYZ instead. The classes get moved out of the basic package
