+++ client/sttp-client4/src/main/scala/sttp/tapir/client/sttp4/ws/WebSocketEndpointToSttpClient.scala
@@ -6,12 +6,12 @@ import sttp.client4.ws.async
 import sttp.model._
 import sttp.tapir._
 import sttp.tapir.client.ClientOutputParams
-import sttp.tapir.client.sttp4.{EndpointToSttpClient, WebSocketEndpointToSttpClientExtensions, WebSocketToPipe}
+import sttp.tapir.client.sttp4.{EndpointToSttpClientBase, WebSocketEndpointToSttpClientExtensions, WebSocketToPipe}
 import sttp.tapir.internal._
 import sttp.ws.WebSocket
 
 private[sttp] class WebSocketEndpointToSttpClient[R <: Streams[_] with WebSockets](wsToPipe: WebSocketToPipe[R])
-    extends EndpointToSttpClient
+    extends EndpointToSttpClientBase
     with WebSocketEndpointToSttpClientExtensions {
   def toSttpRequest[F[_], A, E, O, I](
       e: Endpoint[A, I, E, O, R],
