+++ client/sttp-client4/src/test/scala/sttp/tapir/client/sttp4/SttpClientMultipartTests.scala
@@ -1,8 +1,7 @@
 package sttp.tapir.client.sttp4
 
-import sttp.tapir.client.sttp4.basic.BasicSttpClientTestsSender
 import sttp.tapir.client.tests.ClientMultipartTests
 
-class SttpClientMultipartTests extends BasicSttpClientTestsSender with ClientMultipartTests {
+class SttpClientMultipartTests extends SttpClientTestsSender with ClientMultipartTests {
   multipartTests()
 }
