+++ client/sttp-client4/src/test/scalajvm/sttp/tapir/client/sttp4/SttpClientTestsSender.scala
@@ -1,14 +1,14 @@
-package sttp.tapir.client.sttp4.basic
+package sttp.tapir.client.sttp4
 
 import cats.effect.IO
 import cats.effect.std.Dispatcher
 import sttp.client4._
 import sttp.client4.httpclient.fs2.HttpClientFs2Backend
-import sttp.tapir.client.sttp4.basic.BasicSttpClientInterpreter
+import sttp.tapir.client.sttp4.SttpClientInterpreter
 import sttp.tapir.client.tests.ClientTests
 import sttp.tapir.{DecodeResult, Endpoint}
 
-abstract class BasicSttpClientTestsSender extends ClientTests[Any] {
+abstract class SttpClientTestsSender extends ClientTests[Any] {
   val (dispatcher, closeDispatcher) = Dispatcher.parallel[IO](false).allocated.unsafeRunSync()
   val backend: Backend[IO] = HttpClientFs2Backend[IO](dispatcher).unsafeRunSync()
 
@@ -19,7 +19,7 @@ abstract class BasicSttpClientTestsSender extends ClientTests[Any] {
       args: I,
       scheme: String = "http"
   ): IO[Either[E, O]] = {
-    BasicSttpClientInterpreter()
+    SttpClientInterpreter()
       .toSecureRequestThrowDecodeFailures[A, I, E, O](e, Some(uri"$scheme://localhost:$port"))
       .apply(securityArgs)
       .apply(args)
@@ -33,7 +33,7 @@ abstract class BasicSttpClientTestsSender extends ClientTests[Any] {
       securityArgs: A,
       args: I
   ): IO[DecodeResult[Either[E, O]]] = {
-    BasicSttpClientInterpreter()
+    SttpClientInterpreter()
       .toSecureRequest[A, I, E, O](e, Some(uri"http://localhost:$port"))
       .apply(securityArgs)
       .apply(args)
