+++ client/sttp-client4/src/test/scalajvm/sttp/tapir/client/sttp4/SttpClientZioTests.scala
@@ -2,6 +2,6 @@ package sttp.tapir.client.sttp4
 
 import sttp.tapir.client.tests.ClientBasicTests
 
-class SttpClientSimpleTests extends SttpClientTestsSender with ClientBasicTests {
+class SttpClientZioTests extends SttpClientZioTestsSender with ClientBasicTests {
   tests()
 }
