--- client/sttp-client4/src/test/scalajvm/sttp/tapir/client/sttp4/basic/SttpClientBasicZioTests.scala
@@ -1,7 +1,0 @@
-package sttp.tapir.client.sttp4.basic
-
-import sttp.tapir.client.tests.ClientBasicTests
-
-class SttpClientBasicZioTests extends BasicSttpClientZioTestsSender with ClientBasicTests {
-  tests()
-}
